+++ include/votca/xtp/qmmachine.h
@@ -32,142 +32,12 @@
 #include <votca/xtp/orbitals.h>
 #include <votca/xtp/espfit.h>
 #include <votca/xtp/gdma.h>
-
+#include <votca/xtp/qminterface.h>
+#include <votca/xtp/qmiter.h>
 
 namespace votca { namespace xtp {
 
-    
-// ========================================================================== //
-// QM-MM INTERFACE CLASS - CONVERTS BETWEEN QMATOMS <> POLAR OBJECTS          //
-// ========================================================================== //
-    
-class QMMInterface
-{
-public:
-    
-    QMMInterface() { _polar_table = ctp::POLAR_TABLE(); };
-   ~QMMInterface() {};
-    
-    // CONVERSION QM -> MM
-    ctp::APolarSite *Convert(ctp::QMAtom *atm, int id = -1) {
-        double A_to_nm = 0.1;
-        vec pos = A_to_nm*vec(atm->x, atm->y, atm->z);
-        double q = atm->charge;
-        std::string elem = atm->type;
-        double pol = 0.0;
-        try {
-            pol = _polar_table.at(elem);
-        }
-        catch(out_of_range) {
-            std::cout << std::endl << "QMMInterface - no default polarizability given "
-                << "for element type '" << elem << "'. Defaulting to 1A**3" << std::flush;
-            pol = 1e-3;
-        }
-
-        ctp::APolarSite *new_aps = new ctp::APolarSite(id, elem);
-        new_aps->setRank(0);
-        new_aps->setPos(pos);
-        new_aps->setQ00(q,0); // <- charge state 0 <> 'neutral'
-        new_aps->setIsoP(pol);
-        
-        return new_aps;
-    }
-    
-    ctp::PolarSeg *Convert(std::vector<ctp::QMAtom*> &atms) {        
-        ctp::PolarSeg *new_pseg = new ctp::PolarSeg();
-        std::vector<ctp::QMAtom*>::iterator it;
-        for (it = atms.begin(); it < atms.end(); ++it) {
-            ctp::APolarSite *new_site = this->Convert(*it);
-            new_pseg->push_back(new_site);
-        }
-        return new_pseg;
-    }
-    
-    // TODO CONVERSION MM -> QM
-    ctp::QMAtom *Convert(ctp::APolarSite*);
-    std::vector<ctp::QMAtom*> Convert(ctp::PolarSeg*);
-    
-private:
-    
-    // Allocates polarizabilities in A**3 to element types
-    std::map<std::string,double> _polar_table;
-    
-};
-
-
-
-// ========================================================================== //
-// QM-MM ITERATION CLASS - OBSERVES CONVERGENCE LOOP                          //
-// ========================================================================== //
-    
-class QMMIter
-{
-
-public:
-
-    QMMIter(int id) : _id(id), _hasdRdQ(false), _hasQM(false), _hasMM(false)  { ; }
-   ~QMMIter() { ; }
-
-   void ConvertPSitesToQMAtoms(std::vector< ctp::PolarSeg* > &, std::vector< ctp::QMAtom* > &);
-   void ConvertQMAtomsToPSites(std::vector< ctp::QMAtom* > &, std::vector< ctp::PolarSeg* > &);
-   void UpdatePosChrgFromQMAtoms(std::vector< ctp::QMAtom* > &, std::vector< ctp::PolarSeg* > &);  
-   void UpdateMPSFromGDMA( std::vector<std::vector<double> > &multipoles,  std::vector< ctp::PolarSeg* > &psegs);
-   void GenerateQMAtomsFromPolarSegs(ctp::PolarTop *ptop, Orbitals &orb, bool split_dpl, double dpl_spacing);   
-
-   void setdRdQ(double dR_RMS, double dQ_RMS, double dQ_SUM);
-   void setQMSF(double energy_QM, double energy_SF, double energy_GWBSE);
-   void setE_FM(double ef00, double ef01, double ef02, 
-                  double ef11, double ef12, double em0,
-                  double em1,  double em2, double efm);
-
-   double getRMSdR() { return _dR_RMS; }
-   double getRMSdQ() { return _dQ_RMS; }
-   double getSUMdQ() { return _dQ_SUM; }
-   int getId() { return _id;}
-
-   double getSFEnergy() { assert(_hasQM); return _e_SF; }
-   double getFMEnergy() { assert(_hasMM); return _e_fm_; }
-   double getQMEnergy() { assert(_hasQM); return _e_QM; }
-   double getGWBSEEnergy() { assert(_hasGWBSE); return _e_GWBSE; }
-   double getMMEnergy();
-   double getQMMMEnergy();
-
-
-private:
-
-    int    _id;
-
-    bool   _hasdRdQ;
-    bool   _hasQM;
-    bool   _hasMM;
-    bool   _hasGWBSE;
-
-    double _dR_RMS;
-    double _dQ_RMS;
-    double _dQ_SUM;       
-
-    double _e_QM;
-    double _e_SF;
-    double _e_GWBSE;
-
-    double _ef_00;
-    double _ef_01;
-    double _ef_02;
-    double _ef_11;
-    double _ef_12;
-    double _em_0_;
-    double _em_1_;
-    double _em_2_;
-    double _e_fm_;
-
-
-};
-    
-    
-    
-// ========================================================================== //
-// QMMACHINE: REGISTER QMPACKAGE TYPE (E.G. GAUSSIAN) AT THE END OF .CC FILE  //
-// ========================================================================== //    
+ 
     
 template< class QMPackage >
 class QMMachine
