+++ src/libxtp/qmapemachine.cc
@@ -113,7 +113,7 @@ QMAPEMachine<QMPackage>::QMAPEMachine(ctp::XJob *job, ctp::Ewald3DnD *cape, QMPa
 template<class QMPackage>
 QMAPEMachine<QMPackage>::~QMAPEMachine() {
     
-    std::vector<QMAPEIter*> ::iterator qit;
+    std::vector<QMMIter*> ::iterator qit;
     for (qit = _iters.begin(); qit < _iters.end(); ++qit) {
         delete *qit;
     }
@@ -159,7 +159,7 @@ void QMAPEMachine<QMPackage>::Evaluate(ctp::XJob *job) {
     Orbitals basisforgrid;
     std::vector<ctp::PolarSeg*> dummy;
     
-    GenerateQMAtomsFromPolarSegs(_job->getPolarTop()->QM0(),dummy,basisforgrid);
+ 
     
   
     
@@ -185,7 +185,7 @@ template<class QMPackage>
 bool QMAPEMachine<QMPackage>::Iterate(string jobFolder, int iterCnt) {
 
     // CREATE ITERATION OBJECT & SETUP RUN DIRECTORY
-    QMAPEIter *thisIter = this->CreateNewIter();
+    QMMIter *thisIter = this->CreateNewIter();
     int iter = iterCnt;
     string runFolder = jobFolder + "/iter_" + boost::lexical_cast<string>(iter);
        
@@ -381,9 +381,9 @@ bool QMAPEMachine<QMPackage>::Iterate(string jobFolder, int iterCnt) {
 
 
 template<class QMPackage>
-QMAPEIter *QMAPEMachine<QMPackage>::CreateNewIter() {
+QMMIter *QMAPEMachine<QMPackage>::CreateNewIter() {
     
-    QMAPEIter *newIter = new QMAPEIter(_iters.size());
+    QMMIter *newIter = new QMMIter(_iters.size());
     this->_iters.push_back(newIter);
     return newIter;
 }
@@ -499,17 +499,7 @@ bool QMAPEMachine<QMPackage>::EvaluateGWBSE(Orbitals &orb, string runFolder) {
 		throw runtime_error("Excited state filter yields no states! ");
 
 	}
-	// - output its energy
-        /*
-	double energy___ex = 0.0;
-	if ( _type == "singlet" ){
-		energy___ex = orb.BSESingletEnergies()[_state_index[_state-1]]*13.6058; // to eV
-	} else if ( _type == "triplet" ) {
-		energy___ex = orb.BSETripletEnergies()[_state_index[_state-1]]*13.6058; // to eV
-	}
-*/
-	// ub::matrix<double> &_dft_orbitals_GS = orb_iter_output.MOCoefficients();
-	// int _parse_orbitals_status_GS = _qmpack->ParseOrbitalsFile( &orb_iter_output );
+	
 
 	} // only if state >0
 
@@ -567,8 +557,8 @@ bool QMAPEMachine<QMPackage>::hasConverged() {
     
     if (_iters.size() > 1) {
         
-        QMAPEIter *iter_0 = _iters[_iters.size()-2];
-        QMAPEIter *iter_1 = _iters[_iters.size()-1];
+        QMMIter *iter_0 = _iters[_iters.size()-2];
+        QMMIter *iter_1 = _iters[_iters.size()-1];
         
         double dR = iter_1->getRMSdR();
         double dQ = iter_1->getRMSdQ();
@@ -598,154 +588,6 @@ bool QMAPEMachine<QMPackage>::hasConverged() {
 }
 
 
-void QMAPEIter::ConvertPSitesToQMAtoms(std::vector< ctp::PolarSeg* > &psegs,
-                                       std::vector< ctp::QMAtom * > &qmatoms) {
-    
-    assert(qmatoms.size() == 0);    
-    return;   
-}
-
-
-void QMAPEIter::ConvertQMAtomsToPSites(std::vector< ctp::QMAtom* > &qmatoms,
-                                       std::vector< ctp::PolarSeg* > &psegs) {
-    assert(qmatoms.size() == 0);
-    return;
-}
-
-
-void QMAPEIter::UpdatePosChrgFromQMAtoms(std::vector< ctp::QMAtom* > &qmatoms,
-                                         std::vector< ctp::PolarSeg* > &psegs) {
-    
-    double AA_to_NM = 0.1; // Angstrom to nanometer
-    
-    double dR_RMS = 0.0;
-    double dQ_RMS = 0.0;
-    double dQ_SUM = 0.0;
-    
-    for (unsigned i = 0, qac = 0; i < psegs.size(); ++i) {
-        ctp::PolarSeg *pseg = psegs[i];
-        for (unsigned j = 0; j < pseg->size(); ++j, ++qac) {
-            
-            // Retrieve info from ::QMAtom
-            ctp::QMAtom *qmatm = qmatoms[qac];
-            vec upd_r = vec(qmatm->x, qmatm->y, qmatm->z);
-            upd_r *= AA_to_NM;
-            double upd_Q00 = qmatm->charge;
-            
-            // Compare to previous r, Q00
-            ctp::APolarSite *aps = (*pseg)[j];
-            vec old_r = aps->getPos();
-            double old_Q00 = aps->getQ00();
-            double dR = abs(upd_r - old_r);
-            double dQ00 = upd_Q00 - old_Q00;
-            
-            dR_RMS += dR*dR;
-            dQ_RMS += dQ00*dQ00;
-            dQ_SUM += dQ00;
-            
-            // Forward updated r, Q00 to APS
-            aps->setPos(upd_r);
-            aps->setQ00(upd_Q00, 0);            
-        }
-    }
-    
-    dR_RMS /= qmatoms.size();
-    dQ_RMS /= qmatoms.size();
-    dR_RMS = sqrt(dR_RMS);
-    dQ_RMS = sqrt(dQ_RMS);
-
-    this->setdRdQ(dR_RMS, dQ_RMS, dQ_SUM);
-}
-
-template<class QMPackage>
-void QMAPEMachine<QMPackage>::GenerateQMAtomsFromPolarSegs(std::vector<ctp::PolarSeg*> &qm,
-	std::vector<ctp::PolarSeg*> &mm, Orbitals &orb) {
-    
-    double AA_to_NM = 0.1; // Angstrom to nanometer
-    
-    // QM REGION
-    for (unsigned i = 0; i < qm.size(); ++i) {
-        std::vector<ctp::APolarSite*> *pseg = qm[i];
-        for (unsigned j = 0; j < pseg->size(); ++j) {
-            ctp::APolarSite *aps = (*pseg)[j];
-            string type = "qm";
-            vec pos = aps->getPos()/AA_to_NM;
-            double Q = 0.0;
-            orb.AddAtom(aps->getName(), pos.x(), pos.y(), pos.z(), Q, false);
-        }
-    }
-    
-    // MM REGION (EXPANDED VIA PARTIAL CHARGES)
-    for (unsigned i = 0; i < mm.size(); ++i) {
-    	std::vector<ctp::APolarSite*> *pseg = mm[i];
-        for (unsigned j = 0; j < pseg->size(); ++j) {
-            ctp::APolarSite *aps = (*pseg)[j];
-            string type = "mm";
-            vec pos = aps->getPos()/AA_to_NM;
-            double Q = aps->getQ00();
-            orb.AddAtom(aps->getName(), pos.x(), pos.y(), pos.z(), Q, true);
-        }
-    }
-    
-    return;
-}\
-
-
-
-
-
-void QMAPEIter::setdRdQ(double dR_RMS, double dQ_RMS, double dQ_SUM) {
-    
-    _hasdRdQ = true;    
-    _dR_RMS = dR_RMS;
-    _dQ_RMS = dQ_RMS;
-    _dQ_SUM = dQ_SUM;
-    return;
-}
-
-
-void QMAPEIter::setQMSF(double energy_QM, double energy_SF, double energy_GWBSE) {
-    
-    _hasQM = true;
-    _e_QM = energy_QM;
-    _e_SF = energy_SF;    
-
-    _hasGWBSE = true;
-    _e_GWBSE = energy_GWBSE;
-   
-    return;
-}
-
-
-void QMAPEIter::setE_FM(double ef00, double ef01, double ef02, 
-    double ef11, double ef12, double em0, double em1,  double em2, double efm) {
-    
-    _hasMM = true;
-    _ef_00 = ef00;
-    _ef_01 = ef01;
-    _ef_02 = ef02;
-    _ef_11 = ef11;
-    _ef_12 = ef12;
-    _em_0_ = em0;
-    _em_1_ = em1;
-    _em_2_ = em2;
-    _e_fm_ = efm;
-    return;
-}
-
-
-double QMAPEIter::getMMEnergy() {
-    
-    assert(_hasMM);
-    return _ef_11 + _ef_12 + _em_1_ + _em_2_;
-}
-
-
-double QMAPEIter::getQMMMEnergy() {
-    
-    assert(_hasQM && _hasMM && _hasGWBSE);    
-    return _e_QM + + _e_GWBSE + _ef_11 + _ef_12 + _em_1_ + _em_2_;    
-}
 
 
 // REGISTER QM PACKAGES
